+++ demos/basic/karate/management/commands/loaddemo.py
@@ -1,19 +1,16 @@
 '''
-#---------------------------------------------------------------------------------+
-| Welcome to the swingtime demo project. This project's theme is a Karate dojo    |
-| and the database will be pre-populated with some data relative to today's date. |
-#---------------------------------------------------------------------------------+
+------------------------------------------------------------------------------
+ Welcome to the swingtime demo project. This project's theme is a Karate dojo 
+ and the database will be pre-populated with data relative to today's date.   
+------------------------------------------------------------------------------
 '''
 import os
-import django
+from datetime import datetime, date, time, timedelta
+
 from django.core.management import call_command
 from django.core.management.base import BaseCommand, CommandError
-from django.contrib.auth.models import User
-from datetime import datetime, date, time, timedelta
+from django.contrib.auth import get_user_model
 from django.conf import settings
-from django.db.models import signals
-from django.utils.termcolors import make_style
-from django.core.management.color import color_style
 from dateutil import rrule
 from swingtime import models as swingtime
 
@@ -38,9 +35,9 @@ def create_sample_data():
             ('spc',  'Special Event'),
         )
     ))
-    print(__doc__)
-    print('Created event types: %s' % (
-        ', '.join(['%s' % et for et in swingtime.EventType.objects.all()]),
+    print('{}\nCreated event types: {}'.format(
+        __doc__,
+        ', '.join([str(et) for et in swingtime.EventType.objects.all()]),
     ))
     
     now = datetime.now()
@@ -54,7 +51,7 @@ def create_sample_data():
         end_time=datetime.combine(now.date(), time(18)),
         note='Free tea, sushi, and sake'
     )
-    print('Created event "%s" with %d occurrences' % (evt, evt.occurrence_set.count()))
+    print('Created event "{}" with {} occurrences'.format(evt, evt.occurrence_set.count()))
     
     # create an event with multiple occurrences by fixed count
     evt = swingtime.create_event(
@@ -65,7 +62,7 @@ def create_sample_data():
         count=30,
         byweekday=(rrule.MO, rrule.WE, rrule.FR)
     )
-    print('Created event "%s" with %d occurrences' % (evt, evt.occurrence_set.count()))
+    print('Created event "{}" with {} occurrences'.format(evt, evt.occurrence_set.count()))
 
     # create an event with multiple occurrences by ending date (until)
     evt = swingtime.create_event(
@@ -76,7 +73,7 @@ def create_sample_data():
         until=now + timedelta(days=+70),
         byweekday=(rrule.MO, rrule.WE, rrule.FR)
     )
-    print('Created event "%s" with %d occurrences' % (evt, evt.occurrence_set.count()))
+    print('Created event "{}" with {} occurrences'.format(evt, evt.occurrence_set.count()))
 
     # create an event with multiple occurrences by fixed count on monthly basis
     evt = swingtime.create_event(
@@ -89,7 +86,7 @@ def create_sample_data():
         freq=rrule.MONTHLY,
         byweekday=(rrule.TH(+1), rrule.TH(+3))
     )
-    print('Created event "%s" with %d occurrences' % (evt, evt.occurrence_set.count()))
+    print('Created event "{}" with {} occurrences'.format(evt, evt.occurrence_set.count()))
 
     # create an event with multiple occurrences and alternate intervale
     evt = swingtime.create_event(
@@ -102,18 +99,18 @@ def create_sample_data():
         count=6,
         byweekday=(rrule.SU)
     )
-    print('Created event "%s" with %d occurrences\n' % (evt, evt.occurrence_set.count()))
+    print('Created event "{}" with {} occurrences\n'.format(evt, evt.occurrence_set.count()))
 
 
 class Command(BaseCommand):
+    
     help = 'Run the swingtime demo. If an existing demo database exists, it will recreated.'
     
     def handle(self, **options):
         dbpath = settings.DATABASES['default']['NAME']
         if os.path.exists(dbpath):
-            self.stdout.write(Term.warn('Removing old database %s' % dbpath))
+            self.stdout.write('Removing old database {}'.format(dbpath))
             os.remove(dbpath)
-        self.stdout.write(Term.info('Creating database %s' % dbpath))
 
         call_command('migrate', noinput=True, load_initial_data=False, interactive=False)
         User.objects.create_superuser('admin', 'admin@example.com', 'password')